requirements.txt
anyio==3.6.2
certifi==2024.7.4
chardet==5.1.0
charset-normalizer==3.0.1
gotify==0.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
pyarr==5.2.0
PyYAML==6.0.1
requests==2.32.4
rfc3986==1.5.0
sniffio==1.3.0
tmdbv3api==1.7.7
types-requests==2.28.11.17
types-urllib3==1.26.25.4
urllib3==2.2.2